server: add watchdog and readyness support
This have the service report ready when it successfully published and
subscribed for the first output, and update the watchdog whenever
there's an output update.

As we currently don't dedup these callbacks yet, it should keep the
watchdog happy.

Set Type=notify to ensure it only is considered successfully started
if it discovered at least one display, published state and info, and
subscribed to the set topic.

Add WatchdogSec=10s to ensure an update needs to be processed every 10s,
otherwise it'll also get restarted.